Refactor the utils file.
