+++ src/main/java/sparksoniq/spark/closures/LetClauseMapClosure.java
@@ -25,6 +25,7 @@ import org.apache.spark.sql.types.StructType;
 import sparksoniq.jsoniq.item.Item;
 import sparksoniq.jsoniq.runtime.iterator.RuntimeIterator;
 import sparksoniq.semantics.DynamicContext;
+import sparksoniq.spark.DataFrameUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -61,7 +62,7 @@ public class LetClauseMapClosure implements MapFunction<Row, Row> {
         String[] columnNames = _inputSchema.fieldNames();
 
         // Deserialize row
-        List<Object> deserializedRow = ClosureUtils.deserializeEntireRow(row);
+        List<Object> deserializedRow = DataFrameUtils.deserializeEntireRow(row);
         for (Object columnObject : deserializedRow) {
             List<Item> column = (List<Item>) columnObject;
             _rowColumns.add(column);
@@ -80,6 +81,6 @@ public class LetClauseMapClosure implements MapFunction<Row, Row> {
         }
         _expression.close();
 
-        return ClosureUtils.reserializeRowWithNewData(row, _newColumn, _duplicateColumnIndex);
+        return DataFrameUtils.reserializeRowWithNewData(row, _newColumn, _duplicateColumnIndex);
     }
 }
